alx_travel_app_0x01/alx_travel_app/requirement.txt
django
djangorestframework
django-cors-headers
celery
drfl-yasg
django-environ
mysqlclient
